# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Jarvis Office API
  version: 0.1.0
  description: |
    REST + SSE contract between the Flutter app and the Go api service (backend/cmd/api).

    Conventions
    - Base path `/api/v1`. JSON request/response bodies. Times are RFC 3339 UTC.
    - Money is integer **cents of SGD** in fields ending `_cents` (e.g. 3550 = S$35.50).
    - Fake auth: `POST /auth/login {email}` returns `{token, user}`; send `Authorization: Bearer <token>`
      on every other call. The token is the user id. There is no real security (demo only).
    - Roles: `manager` (create/confirm requests), `approver` (approve/reject), `admin` (everything,
      including `/admin/*`).
    - Errors use `ErrorResponse` with codes: validation_failed (400), unauthorized (401),
      forbidden (403), not_found (404), conflict | invalid_transition | no_active_enrollment (409),
      upstream_error (502), not_implemented (501), internal (500).
    - Request status machine: parsing -> searching -> quoted -> (confirm) -> approved | pending_approval
      -> approved -> checking_out -> awaiting_payment -> paying -> ordered; or rejected / failed / cancelled.
      A live-quote price drift above policy moves checking_out -> pending_approval (kind price_drift).
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      security: []
      summary: Fake login by email (no password)
      description: |
        `{email}` logs in an existing (seeded) user. If the email is unknown or omitted and both `name`
        and `role` are given, a demo user is created on the fly ("any name"). Also sets the
        `jarvis_token` cookie (HttpOnly; value = token), accepted like the Bearer header.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                name: { type: string, description: 'with role: create/login a demo user by name' }
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Current user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Error' }
...
  /api/v1/requests:
    post:
      tags: [requests]
      summary: Create a purchase request (returns immediately in status parsing; progress via SSE)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRequestBody' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PurchaseRequest' }
        '400': { $ref: '#/components/responses/Error' }
    get:
      tags: [requests]
      summary: List requests (newest first)
      parameters:
        - name: status
          in: query
          description: Comma-separated statuses filter
          schema: { type: string, example: 'quoted,pending_approval' }
        - name: mine
          in: query
          schema: { type: boolean }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests: { type: array, items: { $ref: '#/components/schemas/PurchaseRequest' } }
  /api/v1/requests/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [requests]
      summary: Request detail (line items, ranked offers, approvals, payments, address)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequestDetail' }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/requests/{id}/confirm:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [requests]
      summary: Confirm a quoted request for a delivery address
      description: |
        Valid only in status `quoted`. Moves to `approved` (then checkout starts automatically) or to
        `pending_approval`. Returns 409 `no_active_enrollment` if no card is enrolled yet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address_id]
              properties:
                address_id: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequestDetail' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/requests/{id}/cancel:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [requests]
      summary: Cancel (not allowed once paying)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequestDetail' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/requests/{id}/checkout:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [requests]
      summary: Checkout/payment status for a request (from DB; cheap to poll)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutStatus' }
  /api/v1/requests/{id}/checkout/refresh:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [requests]
      summary: Poll Reap GET /agentic/checkouts/:id now for in-flight payments and apply changes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutStatus' }
        '502': { $ref: '#/components/responses/Error' }
  /api/v1/requests/{id}/audit:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [requests]
      summary: Audit trail of one request (oldest first)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: '#/components/schemas/AuditEvent' } }

  # ---------------- events (SSE) ----------------
  /api/v1/events:
    get:
      tags: [events]
      summary: Server-Sent Events stream
      description: |
        `text/event-stream`. Each message has `id: <int>`, `event: <type>` and `data: <SseEvent JSON>`.
        Filter with `request_id`; omit to receive all events (approver console, dashboards).
        Browsers' EventSource cannot set headers, so the token may also be passed as `?token=`.
        A `heartbeat` event is sent every 15 s. Delivery is best-effort; on reconnect re-fetch
        `GET /requests/{id}`.
      parameters:
        - name: request_id
          in: query
          schema: { type: string, format: uuid }
        - name: token
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/SseEvent' }

  # ---------------- approvals ----------------
  /api/v1/approvals:
    get:
      tags: [approvals]
      summary: Approval queue (approver/admin)
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ApprovalStatus' }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  approvals: { type: array, items: { $ref: '#/components/schemas/ApprovalView' } }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/approvals/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [approvals]
      summary: One approval with reasons and the top 3 quotes per line
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovalView' }
  /api/v1/approvals/{id}/approve:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [approvals]
      summary: Approve (approver/admin). The request continues to checkout automatically.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecisionBody' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/approvals/{id}/reject:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    post:
      tags: [approvals]
      summary: Reject (approver/admin). The request moves to rejected.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecisionBody' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }

  # ---------------- orders, spend, audit ----------------
  /api/v1/orders:
    get:
      tags: [orders]
      summary: Order history (requests that reached checkout or later, newest first)
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders: { type: array, items: { $ref: '#/components/schemas/OrderSummary' } }
  /api/v1/spend/monthly:
    get:
      tags: [orders]
      summary: Completed spend per month vs budget (for charts)
      parameters:
        - name: months
          in: query
          schema: { type: integer, default: 6, minimum: 1, maximum: 24 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpendSummary' }
  /api/v1/audit:
    get:
      tags: [orders]
      summary: Global audit log (newest first)
      parameters:
        - name: type
          in: query
          description: Exact type or prefix ending with "." (e.g. "reap.")
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: { type: array, items: { $ref: '#/components/schemas/AuditEvent' } }

  # ---------------- read-only lookups for any role ----------------
  /api/v1/catalog:
    get:
      tags: [requests]
      summary: Active catalog items (read-only)
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/CatalogItem' } }
...
  /api/v1/agent/chat:
    post:
      tags: [agent]
      summary: One text-agent turn (same tools and system prompt as voice)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                session_id: { type: string, description: 'omit to start a new session' }
                message: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatOutput' }
        '502': { $ref: '#/components/responses/Error' }
  /api/v1/agent/sessions/{id}/messages:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [agent]
      summary: Transcript of a text session
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages: { type: array, items: { $ref: '#/components/schemas/ChatMessage' } }
...
  /api/v1/agent/tools/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          enum: [create_order_request, get_request_status, list_addresses, confirm_order, cancel_request]
    post:
      tags: [agent]
      summary: Execute a tool call relayed from the Realtime voice session
      description: |
        The browser receives `response.function_call_arguments.done` from OpenAI Realtime, POSTs the
        arguments here, and sends the returned `output` back as a `function_call_output` item.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [arguments]
              properties:
                call_id: { type: string }
                session_id: { type: string }
                arguments:
                  description: Raw JSON object from the model (object or JSON-encoded string)
                  oneOf:
                    - type: object
                    - type: string
      responses:
        '200':
          description: Tool result (tool-level errors are returned as output.error, still 200)
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_id: { type: string }
                  output: { type: object, additionalProperties: true }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/realtime/session:
    post:
      tags: [agent]
      summary: Mint a short-lived OpenAI Realtime client secret (instructions = DB system prompt)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RealtimeSession' }
        '502': { $ref: '#/components/responses/Error' }

  # ---------------- card enrollment ----------------
  /api/v1/enrollments:
    post:
      tags: [enrollment]
      summary: Start a Reap EXTERNAL card enrollment; open next_action_url (hosted card-entry page)
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '502': { $ref: '#/components/responses/Error' }
  /api/v1/enrollments/current:
    get:
      tags: [enrollment]
      summary: Latest enrollment, refreshed from Reap while not terminal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '404': { $ref: '#/components/responses/Error' }

  # ---------------- admin (role admin) ----------------
  /api/v1/admin/catalog:
    get:
      tags: [admin]
      summary: All catalog items (including inactive)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/CatalogItem' } }
    post:
      tags: [admin]
      summary: Create catalog item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CatalogItemInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogItem' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/admin/catalog/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogItem' }
    put:
      tags: [admin]
      summary: Replace catalog item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CatalogItemInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogItem' }
    delete:
      tags: [admin]
      responses:
        '204': { description: Deleted }
...
  /api/v1/admin/policy:
    get:
      tags: [admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyConfig' }
    put:
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PolicyInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyConfig' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Fake token = user id from POST /auth/login
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    Role:
      type: string
      enum: [manager, approver, admin]
...
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
...